+++ src/utils/authOptions.ts
@@ -5,6 +5,10 @@ import { User } from "@prisma/client";
 import { prisma } from "@/lib/db";
 import { revalidatePath } from "next/cache";
 
+const GOOGLE_CLIENT_ID = process.env.GOOGLE_CLIENT_ID!;
+const GOOGLE_CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET!;
+const secret = process.env.NEXTAUTH_SECRET;
+
 export const authOptions: NextAuthOptions = {
     session: {
         strategy: "jwt",
